create BookRecords class
