+++ main.rb
@@ -3,15 +3,15 @@ require_relative 'app'
 def start_app(app, choice)
   case choice
   when 1
-    app.list_books
+    app.books.list_books
   when 2
     app.list_people
   when 3
     app.create_person
   when 4
-    app.create_book
+    app.books.create_book
   when 5
-    app.rentals.create_rental(app.books, app.people)
+    app.rentals.create_rental(app.books.list, app.people)
   when 6
     app.rentals.list_rentals(app.people)
   else
